requirements.txt
streamlit>=1.28.0
pymongo>=4.5.0
python-dotenv>=1.0.0
sib-api-v3-sdk>=7.6.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
pandas>=2.0.0
plotly>=5.17.0
python-dateutil>=2.8.2